Tac class now a std::variant<std::string,int> for LHS and RHS operands.
This will make it easier to generate bytecode for our IR.
For example, if some operation has one (or two) integer operands, then
those operands must be pushed to the stack using the "iconst" instruction.
But if an operand is a variable, then we must instead use the "iload" instruction.
By storing our operands as a variant, we can handle this without resorting
to checking if an operand is an integer, which is ugly and would probably
lead to poor performance.